Add test for cleartext signatures
Currently, simple signatures are expected to be in the binary GPG
format, and that's what e.g. `podman push --sign-by` produces as well.
But in fact, the code as is also works with the cleartext signatures.
E.g. for the gpgme backend, the same `_verify` API can handle both
binary and cleartext signatures.

Add a new test to cover this case. This then is implying that this is a
supported use case.

The reason why I'm interested in this is that I'd like to make use
of it for signing Fedora CoreOS container images. The end-goal is to
move to Sigstore signing, but until that's ready, we'd like to use GPG
signing.

We use Robosignatory, the Fedora signing service, which only supports
detached signatures, and while it's theoretically possible to convert
the detached signatures we get back into inline binary signatures,
it's much less cumbersome and error-prone to convert it to cleartext
signatures.

It's worth noting that while Fedora's signing server (Sigul) does
support container image signing, Robosignatory does not surface it yet
(see https://pagure.io/robosignatory/issue/22).

Fixing that wouldn't be too hard I think, but all this is
ideally short-term anyway until we can move to Sigstore
signing + Konflux. There's work in progress on that (see e.g.
https://discussion.fedoraproject.org/t/148999).